refactor: split AI account status modules
